fix component tests: remove unused jest config, rename tests to .test instead of .spec
and updated the snapshot of components
