--- src/components/Dropdown/__tests__/Dropdown.options.ts
@@ -1,13 +1,0 @@
-import { DropdownOption } from "..";
-
-export const options: DropdownOption[] = [
-    {value: "You haven't selected a name yet!", type: "disabled"},
-    {value: "Karen", type: "header"},
-    {type: "divider"},
-    {value: "Gillian", type: "header"},
-    {type: "divider"},
-    {value: "Bucky", type: "header"},
-    {type: "divider"},
-    {value: "Bikrum", type: "header"},
-    {type: "divider"}
-]
