+++ src/components/Label/index.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import { Label as RSLabel } from 'reactstrap';
 import { LabelProps, LabelObject } from '../../common/interfaces/CommonProps';
 export const Label: React.FunctionComponent<LabelProps> = function Label(props) {
-    const { value, className, onClick } = props;
+    const { value, className, onClick, noLabel } = props;
+    if(noLabel || !value) return null;
     const labelInfo = typeof value === typeof "" ?
         { labelFor: value as string, labelValue: value as string } :
             { labelFor: (value as LabelObject).htmlFor, labelValue: (value as LabelObject).value };
